net: switch packet classifier to associate packets with rings, not sockets
This is a step on the way to solve issues with socket lifetimes.

When destroying a socket, we must remove it from the classifier.  However,
there is no shared lock between the classifier and the socket, so we
can't just remove it, and we can't use message passing and wait for an ack,
as that would cause deadlocks.

As a first step, have the classifier talk to a ring, not a socket.  The next
patch will shift ownership of the ring to the classifier.

Signed-off-by: Avi Kivity <avi@cloudius-systems.com>