sched: support for waiting for multiple objects of different types
Waiting is complicated; you can wait for a condition variable, a timer,
wait queues and signals (when we have them), semaphores, and more.

When you need to wait for multiple events, things become even more
complicated as some sort of abstraction is needed.  poll() for example
supports waiting for multiple objects that conform to the fd interface,
or for a timeout.

To allow more flexible waiting, this patch adds a poll-like interface
to the scheduler, except that the interfaces are all compile time.

We add a new function wait_for() that accepts a variable number of waitable
objects.  A waitable object is an object that supports the three methods:

 - poll(): check if waiting is complete
 - arm(): prepare for waiting
 - disarm(): finish waiting

to make a waitable object usable with wait_for, these functions have to be
supplied.

Signed-off-by: Avi Kivity <avi@cloudius-systems.com>