rcu_list: add splice_front()
This is useful when we want to add an element to a list, but the list that
is being modified depends on the element itself, and the element has not
been constructed yet (e.g. we're emplacing the element, not copying an existing
one).

The solution is splicing: we emplace the element into a temporary list,
consult it to determine which list it needs to be placed in, and finally
splice the temporary list into the destination list.

Signed-off-by: Avi Kivity <avi@cloudius-systems.com>