net: fix net channel lookup of removed channel
Under a heavy wrk workload with many connections, we sometimes lookup
a net channel which has already been deleted.  This is due to ports being
reused, so we have two connections with the same key in the hash table.
When the old one is deleted, since we only match the tuple, we may actually
remove the new one; when we next look it up we may find the deleted channel.

This likely indicates that freeing is delayed too much, but for now, it is
better to fix it by matching against both id and channel, than crashing.

Signed-off-by: Avi Kivity <avi@cloudius-systems.com>